CODE: Don't use long. possible 64bit multiplayer grief (TurfIt, from Standard)

+++ gui/components/gui_textinput.h
@@ -75,7 +75,7 @@ protected:
 	 * reference time for regulating cursor blinking
 	 * @author Knightly
 	 */
-	unsigned long cursor_reference_time;
+	uint32 cursor_reference_time;
 
 	/**
 	 * whether focus has been received
